moved fetch call from useEffect in GetItinerary to onSubmit in MainMenu
